+++ crates/utils/src/lib.rs
@@ -4,6 +4,7 @@ pub mod file;
 pub mod imap;
 pub mod metadata;
 pub mod mime;
+pub mod storage;
 pub mod tls;
 
 pub use auth::*;
@@ -12,4 +13,5 @@ pub use file::*;
 pub use imap::*;
 pub use metadata::*;
 pub use mime::*;
+pub use storage::*;
 pub use tls::*;